+++ src/main/java/io/airlift/tpch/TpchColumnTypes.java
@@ -20,7 +20,7 @@ public class TpchColumnTypes
     public static final TpchColumnType INTEGER = new TpchColumnType(TpchColumnType.Base.INTEGER);
     public static final TpchColumnType IDENTIFIER = new TpchColumnType(TpchColumnType.Base.IDENTIFIER);
     public static final TpchColumnType DATE = new TpchColumnType(TpchColumnType.Base.DATE);
-    public static final TpchColumnType DOUBLE = new TpchColumnType(TpchColumnType.Base.DOUBLE);
+    public static final TpchColumnType DECIMAL = new TpchColumnType(TpchColumnType.Base.DECIMAL);
 
     public static TpchColumnType varchar(long precision) {
         return new TpchColumnType(TpchColumnType.Base.VARCHAR, precision);
